APImarket/requirements.txt
aerich==0.7.2
aiofiles==23.2.1
aiohttp==3.8.5
aiomysql==0.2.0
aiosignal==1.3.1
aiosqlite==0.17.0
aliyun-python-sdk-core==2.13.36
annotated-types==0.5.0
anyio==3.7.1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bilibili-api-python==15.5.5
Brotli==1.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
dictdiffer==0.9.0
frozenlist==1.4.0
h11==0.14.0
html5tagger==1.3.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
iso8601==1.1.0
jmespath==0.10.0
lxml==4.9.3
multidict==6.0.4
passlib==1.7.4
Pillow==10.0.0
pyasn1==0.5.0
pycparser==2.21
pycryptodomex==3.18.0
pydantic==2.3.0
pydantic_core==2.6.3
PyMySQL==1.1.0
pypika-tortoise==0.1.6
pypng==0.20220715.0
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
qrcode-terminal==0.8
redis==5.0.0
requests==2.31.0
rsa==4.9
sanic==23.6.0
sanic-ext==23.6.0
sanic-routing==23.6.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
tomlkit==0.12.1
tortoise-orm==0.20.0
tqdm==4.66.1
tracerite==1.1.0
typing_extensions==4.7.1
tzlocal==5.0.1
ua-parser==0.18.0
ujson==5.8.0
urllib3==2.0.4
user-agents==2.2.0
uvloop==0.17.0
websockets==11.0.3
xhs==0.2.5
yarl==1.9.2
